# Adds database session caching