This should radically improve performance under heavy load.

**Note**: the page load times below are representative instances of the typical page load times for the given request. I repeated the requests many times, and chose examples that were near the median value of the sample set.

# Loading Attendee Registration Details

### Without Caching
* 12 new sessions
* Page loads in ~0.2 secs

```
NEW SESSION: <sqlalchemy.orm.session.Session object at 0x7f2508224828>
    NEW SESSION: <sqlalchemy.orm.session.Session object at 0x7f252864ceb8>
    CLOSING SESSION: <sqlalchemy.orm.session.Session object at 0x7f252864ceb8>
    NEW SESSION: <sqlalchemy.orm.session.Session object at 0x7f252864c358>
        NEW SESSION: <sqlalchemy.orm.session.Session object at 0x7f25280afba8>
            NEW SESSION: <sqlalchemy.orm.session.Session object at 0x7f25082c7cf8>
            CLOSING SESSION: <sqlalchemy.orm.session.Session object at 0x7f25082c7cf8>
        CLOSING SESSION: <sqlalchemy.orm.session.Session object at 0x7f25280afba8>
    CLOSING SESSION: <sqlalchemy.orm.session.Session object at 0x7f252864c358>
    NEW SESSION: <sqlalchemy.orm.session.Session object at 0x7f25082b3be0>
    CLOSING SESSION: <sqlalchemy.orm.session.Session object at 0x7f25082b3be0>
    NEW SESSION: <sqlalchemy.orm.session.Session object at 0x7f25082c7358>
    CLOSING SESSION: <sqlalchemy.orm.session.Session object at 0x7f25082c7358>
    NEW SESSION: <sqlalchemy.orm.session.Session object at 0x7f25082b3518>
    CLOSING SESSION: <sqlalchemy.orm.session.Session object at 0x7f25082b3518>
    NEW SESSION: <sqlalchemy.orm.session.Session object at 0x7f2528094be0>
    CLOSING SESSION: <sqlalchemy.orm.session.Session object at 0x7f2528094be0>
    NEW SESSION: <sqlalchemy.orm.session.Session object at 0x7f25280bf8d0>
    CLOSING SESSION: <sqlalchemy.orm.session.Session object at 0x7f25280bf8d0>
    NEW SESSION: <sqlalchemy.orm.session.Session object at 0x7f25082b30f0>
    CLOSING SESSION: <sqlalchemy.orm.session.Session object at 0x7f25082b30f0>
    NEW SESSION: <sqlalchemy.orm.session.Session object at 0x7f2528037d68>
    CLOSING SESSION: <sqlalchemy.orm.session.Session object at 0x7f2528037d68>
CLOSING SESSION: <sqlalchemy.orm.session.Session object at 0x7f2508224828>
2018-08-13 03:35:22,953 [DEBUG] uber.decorators: uber.site_sections.registration.form loaded in 0.212305 seconds
2018-08-13 03:35:22,954 [INFO] cherrypy.access.139798003845384: 10.0.2.2 - - [13/Aug/2018:03:35:22] "GET /reggie/registration/form?id=aec400a8-283e-4afe-a330-f0d8c162417f HTTP/1.1" 200 72596 "https://localhost:4443/registration/" "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_13_6) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/11.1.2 Safari/605.1.15"
```

### With Caching
* 1 new session
* Page loads in ~0.1 sec

```
NEW SESSION: <sqlalchemy.orm.session.Session object at 0x7f2d220d4b38>
    CACHED SESSION: <sqlalchemy.orm.session.Session object at 0x7f2d220d4b38>
    CACHED SESSION: <sqlalchemy.orm.session.Session object at 0x7f2d220d4b38>
    CACHED SESSION: <sqlalchemy.orm.session.Session object at 0x7f2d220d4b38>
    CACHED SESSION: <sqlalchemy.orm.session.Session object at 0x7f2d220d4b38>
    CACHED SESSION: <sqlalchemy.orm.session.Session object at 0x7f2d220d4b38>
    CACHED SESSION: <sqlalchemy.orm.session.Session object at 0x7f2d220d4b38>
    CACHED SESSION: <sqlalchemy.orm.session.Session object at 0x7f2d220d4b38>
    CACHED SESSION: <sqlalchemy.orm.session.Session object at 0x7f2d220d4b38>
    CACHED SESSION: <sqlalchemy.orm.session.Session object at 0x7f2d220d4b38>
    CACHED SESSION: <sqlalchemy.orm.session.Session object at 0x7f2d220d4b38>
    CACHED SESSION: <sqlalchemy.orm.session.Session object at 0x7f2d220d4b38>
CLOSING SESSION: <sqlalchemy.orm.session.Session object at 0x7f2d220d4b38>
2018-08-13 03:36:34,818 [DEBUG] uber.decorators: uber.site_sections.registration.form loaded in 0.094528 seconds
2018-08-13 03:36:34,820 [INFO] cherrypy.access.139834081298584: 10.0.2.2 - - [13/Aug/2018:03:36:34] "GET /reggie/registration/form?id=aec400a8-283e-4afe-a330-f0d8c162417f HTTP/1.1" 200 72596 "https://localhost:4443/registration/" "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_13_6) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/11.1.2 Safari/605.1.15"
```


# Loading Preregistration Page

### Without Caching

* 3 new sessions
* Page loads in ~ 0.35 secs

```
NEW SESSION: <sqlalchemy.orm.session.Session object at 0x7f452803d7f0>
    NEW SESSION: <sqlalchemy.orm.session.Session object at 0x7f4518167ef0>
    CLOSING SESSION: <sqlalchemy.orm.session.Session object at 0x7f4518167ef0>
    NEW SESSION: <sqlalchemy.orm.session.Session object at 0x7f44f44cc588>
    CLOSING SESSION: <sqlalchemy.orm.session.Session object at 0x7f44f44cc588>
CLOSING SESSION: <sqlalchemy.orm.session.Session object at 0x7f452803d7f0>
2018-08-13 03:41:16,165 [DEBUG] uber.decorators: uber.site_sections.preregistration.form loaded in 0.364221 seconds
2018-08-13 03:41:16,166 [INFO] cherrypy.access.139935184459928: 10.0.2.2 - - [13/Aug/2018:03:41:16] "GET /reggie/preregistration/form HTTP/1.1" 200 66955 "https://localhost:4443/common/" "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_13_6) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/11.1.2 Safari/605.1.15"
```

### With Caching

* 1 new sessions
* Page loads in ~0.2 secs

```
NEW SESSION: <sqlalchemy.orm.session.Session object at 0x7f2d21cbd208>
    CACHED SESSION: <sqlalchemy.orm.session.Session object at 0x7f2d21cbd208>
    CACHED SESSION: <sqlalchemy.orm.session.Session object at 0x7f2d21cbd208>
CLOSING SESSION: <sqlalchemy.orm.session.Session object at 0x7f2d21cbd208>
2018-08-13 03:39:53,949 [DEBUG] uber.decorators: uber.site_sections.preregistration.form loaded in 0.199263 seconds
2018-08-13 03:39:53,951 [INFO] cherrypy.access.139834081298584: 10.0.2.2 - - [13/Aug/2018:03:39:53] "GET /reggie/preregistration/form HTTP/1.1" 200 66955 "https://localhost:4443/common/" "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_13_6) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/11.1.2 Safari/605.1.15"
```